Coding Jan 1st
Added more functionality to the game. Ghost system was added so the AI isn't a heat seeking missle (lol). Also more reasonable towards human behaviour. Added some more code though, which means I need to access the ISANGLE, ISDETECTED from the stealth script and make it a requirement for the time ticking down as that's the only current way for a game over. Tomorrow is UI and implementing a finished level prototype.